third_party/WebKit/LayoutTests/custom-properties/usecount-register-property.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let CSSRegisterProperty = 2018; // From UseCounter.h

    let isCounted = () => internals.isUseCounted(document, CSSRegisterProperty);

    assert_false(isCounted(),
                 'should not be counted before registerProperty call');

    CSS.registerProperty({name: '--name1'});

    assert_true(isCounted(),
                'registerProperty call should be counted');
}, "registerProperty is use counted");
</script>
